feat: add the experimental parser to the `wdl-grammar` binary.
This commit alters the `wdl-grammar` binary to use the new parser
implementation depending on the enabling of the `experimental` feature.

It keeps the existing implementation of the binary in place when the feature is
off.

Currently it only accepts a `parse` command with the path to the WDL file to
parse or `-` for reading from stdin.

The `parse` command prints any parser errors it encounters as well as the
resulting syntax tree.

Use the following to run the CLI:

```
cargo run --release --features binaries,experimental --bin wdl-grammar -- parse
$FILE
```